Disable future checking only in --raft mode
It's crucial to disable future checking in --raft mode because block validation
in the raft setting needs to be deterministic. There is no forking of the
chain, and we need each node to only perform validation as a pure function of
block contents with respect to the previous database state.

Whereas we are currently checking whether the timestamp field has nanosecond
semantics to detect --raft mode, we could instead use a special "raft" sentinel
in the Extra field, or pass a boolean for raftMode from all call sites of this
function.